backend/requirements.txt
# ────────────────────────────────────────────────────────────────────────────────
# Plant Disease AI — Python Dependencies
# Pin major.minor, allow patch updates for security fixes
# ────────────────────────────────────────────────────────────────────────────────

# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi==0.110.*
uvicorn[standard]==0.29.*
# uvloop==0.19.*
httptools==0.6.*
python-multipart==0.0.9       # file uploads

# ── Auth & Security ────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
cryptography==42.*

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.*
asyncpg==0.29.*               # async PostgreSQL driver
alembic==1.13.*               # migrations
motor==3.4.*                   # async MongoDB driver
pymongo==4.7.*

# ── Caching / Queue ───────────────────────────────────────────────────────────
redis==5.0.*
celery[redis]==5.3.*
flower==2.0.*

# ── Object Storage ─────────────────────────────────────────────────────────────
aioboto3==12.*
boto3==1.34.*

# ── ML / Deep Learning ────────────────────────────────────────────────────────
# NOTE: For GPU, replace with:
#   torch==2.2.*+cu121 --extra-index-url https://download.pytorch.org/whl/cu121
torch==2.2.*
torchvision==0.17.*
torchaudio==2.2.*

# ── Image Processing ──────────────────────────────────────────────────────────
opencv-python-headless==4.9.*  # headless — no GUI deps
Pillow==10.*
scikit-image==0.23.*

# ── Scientific / ML Utils ──────────────────────────────────────────────────────
numpy==1.26.*
scipy==1.13.*
scikit-learn==1.4.*
pandas==2.2.*

# ── MLflow ────────────────────────────────────────────────────────────────────
mlflow==2.11.*

# ── Observability ─────────────────────────────────────────────────────────────
structlog==24.*
sentry-sdk[fastapi]==1.45.*
prometheus-fastapi-instrumentator==6.*
opentelemetry-api==1.24.*
opentelemetry-sdk==1.24.*

# ── Config ────────────────────────────────────────────────────────────────────
pydantic==2.7.*
pydantic-settings==2.2.*

# ── HTTP Client (for advisory service, external APIs) ─────────────────────────
httpx==0.27.*
aiohttp==3.9.*

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.*
tenacity==8.2.*                # retries
orjson==3.10.*                 # fast JSON
tqdm==4.66.*                   # progress bars

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.*
pytest-asyncio==0.23.*
pytest-cov==5.*
httpx==0.27.*                  # async test client
factory-boy==3.3.*             # test fixtures
